requirements.txt
fastapi
fastapi_pagination
jose
FLASK_APP
pddb-genwch>=0.0.5
uvicorn